refactor: move blog to blog/index and add typescript path aliases
