+++ src/pages/404.astro
@@ -1,5 +1,5 @@
 ---
-import Layout from '../layouts/layout.astro';
+import Layout from '@layouts/layout.astro';
 ---
 
 <Layout>
