+++ src/pages/blog/index.astro
@@ -1,7 +1,7 @@
 ---
 import { getCollection } from 'astro:content';
-import Layout from '../layouts/layout.astro';
-import { readingTimeKey } from '../util/remark-reading-time';
+import Layout from '@layouts/layout.astro';
+import { readingTimeKey } from '@utils/remark-reading-time';
 
 const posts = await getCollection('blog');
 const filteredPosts = posts.filter((post) => import.meta.env.DEV || !post.data.draft);
