+++ src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-import Layout from '../layouts/layout.astro';
+import Layout from '@layouts/layout.astro';
 ---
 
 <Layout title="Home" description="Homepage of Joep Kockelkorn, full-stack Developer && front-end fanatic.">
